feat: Tuning unit tests and correct access to property values
