+++ integrationtest/src/test/java/org/mapstruct/itest/tests/FullFeatureCompilationExclusionCliEnhancer.java
@@ -31,7 +31,7 @@ public final class FullFeatureCompilationExclusionCliEnhancer implements Process
             case JAVA_8:
                 additionalExcludes.add( "org/mapstruct/ap/test/injectionstrategy/cdi/**/*.java" );
                 additionalExcludes.add( "org/mapstruct/ap/test/injectionstrategy/jakarta_cdi/**/*.java" );
-                additionalExcludes.add( "org/mapstruct/ap/test/annotatewith/deprecated/newer/*.java" );
+                additionalExcludes.add( "org/mapstruct/ap/test/annotatewith/deprecated/jdk11/*.java" );
                 break;
             case JAVA_9:
                 // TODO find out why this fails:
