+++ processor/src/test/java/org/mapstruct/ap/test/annotatewith/deprecated/DeprecatedMapperWithClass.java
@@ -3,13 +3,11 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.annotatewith.deprecated.older;
+package org.mapstruct.ap.test.annotatewith.deprecated;
 
-import org.mapstruct.AnnotateWith;
 import org.mapstruct.Mapper;
 
 @Mapper
 @Deprecated
-@AnnotateWith(Deprecated.class)
-public class OlderRepeatDeprecatedMapper {
+public class DeprecatedMapperWithClass {
 }
