+++ processor/src/test/java/org/mapstruct/ap/test/annotatewith/deprecated/DeprecatedMapperWithMethod.java
@@ -3,14 +3,14 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.annotatewith.deprecated.older;
+package org.mapstruct.ap.test.annotatewith.deprecated;
 
 import org.mapstruct.AnnotateWith;
 import org.mapstruct.Mapper;
 import org.mapstruct.ap.test.annotatewith.CustomMethodOnlyAnnotation;
 
 @Mapper
-public interface OlderDeprecatedMapperWithMethod {
+public interface DeprecatedMapperWithMethod {
 
     @AnnotateWith(CustomMethodOnlyAnnotation.class)
     @Deprecated
