+++ processor/src/test/java/org/mapstruct/ap/test/annotatewith/deprecated/RepeatDeprecatedMapper.java
@@ -3,11 +3,13 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.annotatewith.deprecated.newer;
+package org.mapstruct.ap.test.annotatewith.deprecated;
 
+import org.mapstruct.AnnotateWith;
 import org.mapstruct.Mapper;
 
 @Mapper
-@Deprecated(since = "11")
-public class NewerDeprecatedMapperWithClass {
+@Deprecated
+@AnnotateWith(Deprecated.class)
+public class RepeatDeprecatedMapper {
 }
