+++ processor/src/test/java/org/mapstruct/ap/test/annotatewith/deprecated/jdk11/DeprecatedMapperWithClass.java
@@ -3,11 +3,11 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.annotatewith.deprecated.older;
+package org.mapstruct.ap.test.annotatewith.deprecated.jdk11;
 
 import org.mapstruct.Mapper;
 
 @Mapper
-@Deprecated
-public class OlderDeprecatedMapperWithClass {
+@Deprecated(since = "11")
+public class DeprecatedMapperWithClass {
 }
