+++ processor/src/test/java/org/mapstruct/ap/test/annotatewith/deprecated/jdk11/RepeatDeprecatedMapperWithParams.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.annotatewith.deprecated.newer;
+package org.mapstruct.ap.test.annotatewith.deprecated.jdk11;
 
 import org.mapstruct.AnnotateWith;
 import org.mapstruct.Mapper;
@@ -17,5 +17,5 @@ import org.mapstruct.Mapper;
         @AnnotateWith.Element( name = "since", strings = "1.5")
     }
 )
-public class NewerRepeatDeprecatedMapperWithParams {
+public class RepeatDeprecatedMapperWithParams {
 }
