--- processor/src/test/java/org/mapstruct/ap/test/annotatewith/deprecated/newer/NewerRepeatDeprecatedMapper.java
@@ -1,15 +1,0 @@
-/*
- * Copyright MapStruct Authors.
- *
- * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
- */
-package org.mapstruct.ap.test.annotatewith.deprecated.newer;
-
-import org.mapstruct.AnnotateWith;
-import org.mapstruct.Mapper;
-
-@Mapper
-@Deprecated
-@AnnotateWith(Deprecated.class)
-public class NewerRepeatDeprecatedMapper {
-}
